Updated sway, waybar config files.

+++ home/.config/sway/script/close_window.sh
@@ -2,10 +2,16 @@
 
 set -euo pipefail
 
+# Close the current focusing window.
 # swaymsg kill
-swaynag \
-    -t warning \
-    -m 'Do you want to close the current focusing window?' \
-    -b 'Yes, close window' 'swaymsg kill' \
-    -s "No / Close this notification"
-pkill swaynag
+
+if ! pgrep swaynag > /dev/null; then
+    # Print a warning message to close the window.
+    swaynag \
+        -t warning \
+        -m 'Do you want to close the current focusing window?' \
+        -b 'Yes, close window' 'swaymsg kill' \
+        -s "No / Close this notification"
+else
+    pkill swaynag
+fi
